+++ tests/test_common_super_resolution_format.py
@@ -8,19 +8,19 @@ from datumaro.components.dataset import Dataset
 from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Image
-from datumaro.plugins.super_resolution_format import SuperResolutionImporter
+from datumaro.plugins.common_super_resolution_format import CommonSuperResolutionImporter
 from datumaro.util.test_utils import compare_datasets
 
 from .requirements import Requirements, mark_requirement
 
-DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), "assets", "super_resolution_dataset")
+DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), "assets", "common_super_resolution_dataset")
 
 
-class SuperResolutionImporterTest(TestCase):
+class CommonSuperResolutionImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_detect(self):
         detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
-        self.assertEqual([SuperResolutionImporter.NAME], detected_formats)
+        self.assertEqual([CommonSuperResolutionImporter.NAME], detected_formats)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_import(self):
@@ -49,6 +49,6 @@ class SuperResolutionImporterTest(TestCase):
             ]
         )
 
-        dataset = Dataset.import_from(DUMMY_DATASET_DIR, "super_resolution")
+        dataset = Dataset.import_from(DUMMY_DATASET_DIR, "common_super_resolution")
 
         compare_datasets(self, expected_dataset, dataset, require_media=True)